test(e2e): Port `deno` E2E test app to span streaming
Removes the `traceLifecycle: 'static'` pin from the `deno` app and
rewrites its specs against streamed spans. Adds `deno-static` as a
verbatim copy of the pre-port app to keep the static trace lifecycle
covered, and removes `deno-streamed`, which is now a strict subset of
`deno`.
