+++ .github/workflows/build.yml
@@ -1041,7 +1041,7 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
       - name: Set up Deno
         if:
-          matrix.test-application == 'deno' || matrix.test-application == 'deno-streamed' || matrix.test-application ==
+          matrix.test-application == 'deno' || matrix.test-application == 'deno-static' || matrix.test-application ==
           'deno-redis' || matrix.test-application == 'hono-4' || matrix.test-application == 'deno-mysql' ||
           matrix.test-application == 'deno-pg'
         uses: denoland/setup-deno@v2.0.5
@@ -1166,7 +1166,7 @@ jobs:
         with:
           node-version-file: 'dev-packages/e2e-tests/test-applications/${{ matrix.test-application }}/package.json'
       - name: Set up Deno
-        if: matrix.test-application == 'deno' || matrix.test-application == 'deno-streamed'
+        if: matrix.test-application == 'deno'
         uses: denoland/setup-deno@v2.0.5
         with:
           deno-version: ${{ matrix.deno-version || 'v2.8.3' }}
