+++ dev-packages/e2e-tests/test-applications/deno-static/deno.json
