+++ dev-packages/e2e-tests/test-applications/deno-static/playwright.config.mjs
