+++ dev-packages/e2e-tests/test-applications/deno-static/src/app.ts
@@ -18,6 +18,7 @@ import { MockLanguageModelV1 } from 'ai/test';
 import { z } from 'zod';
 
 Sentry.init({
+  traceLifecycle: 'static',
   environment: 'qa',
   dsn: Deno.env.get('E2E_TEST_DSN'),
   debug: !!Deno.env.get('DEBUG'),
