+++ dev-packages/e2e-tests/test-applications/deno/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "deno run --allow-net --allow-env --allow-read src/app.ts",
+    "start": "deno run --allow-net --allow-env --allow-read --allow-sys src/app.ts",
     "test": "playwright test",
     "clean": "npx rimraf node_modules pnpm-lock.yaml",
     "test:build": "pnpm install",
